2014-03-04  Bill Schmidt  <wschmidt@linux.vnet.ibm.com>
	* gcc.dg/vmx/extract-vsx.c: Replace "vector long" with "vector
	long long" throughout.
	* gcc.dg/vmx/extract-vsx-be-order.c: Likewise.
	* gcc.dg/vmx/insert-vsx.c: Likewise.
	* gcc.dg/vmx/insert-vsx-be-order.c: Likewise.
	* gcc.dg/vmx/ld-vsx.c: Likewise.
	* gcc.dg/vmx/ld-vsx-be-order.c: Likewise.
	* gcc.dg/vmx/ldl-vsx.c: Likewise.
	* gcc.dg/vmx/ldl-vsx-be-order.c: Likewise.
	* gcc.dg/vmx/merge-vsx.c: Likewise.
	* gcc.dg/vmx/merge-vsx-be-order.c: Likewise.
	* gcc.dg/vmx/st-vsx.c: Likewise.
	* gcc.dg/vmx/st-vsx-be-order.c: Likewise.
	* gcc.dg/vmx/stl-vsx.c: Likewise.
	* gcc.dg/vmx/stl-vsx-be-order.c: Likewise.



git-svn-id: svn+ssh://gcc.gnu.org/svn/gcc/trunk@208321 138bc75d-0d04-0410-961f-82ee72b054a4

+++ gcc/testsuite/gcc.dg/vmx/extract-vsx-be-order.c
@@ -6,7 +6,7 @@
 
 static void test()
 {
-  vector long vl = {0, 1};
+  vector long long vl = {0, 1};
   vector double vd = {0.0, 1.0};
 
 #if __BYTE_ORDER__ == __ORDER_LITTLE_ENDIAN__
